[python] refactored tests and import shared utilities from module
